requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
Pillow
streamlit